backend/requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
joblib==1.3.2
sqlalchemy==2.0.20
fastapi==0.103.1
uvicorn==0.23.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.3.0
bcrypt==4.0.1 